feat(my-page): mark-up notification-email
