+++ packages/language-pack/ja.json
@@ -36,6 +36,12 @@
 				"TITLE": "",
 				"USER_ID": ""
 			},
+			"MOBILE": {
+				"MOBILE_GUIDE_MODAL_DESC_1": "",
+				"MOBILE_GUIDE_MODAL_DESC_2": "",
+				"MOBILE_GUIDE_MODAL_DESC_3": "",
+				"MOBILE_GUIDE_MODAL_TITLE": ""
+			},
 			"RESET": {
 				"CONFIRM_PASSWORD": "",
 				"DESC": "",
@@ -71,18 +77,6 @@
 				"RESET_PASSWORD": "パスワード再設定",
 				"TITLE": ""
 			}
-		},
-		"RESET_PASSWORD_PAGE": {
-			"ALREADY_RESET_TEXT": "パスワードが既に変更されました。変更されたパスワードでログインしください。",
-			"GO_TO_LOGIN_PAGE": "ログインページへ移動",
-			"GO_TO_LOGIN_WITHOUT_CHANGE": "パスワードを変更せずにログインに戻る",
-			"HELP_TEXT": "アカウントの新しいパスワード入力:",
-			"INVALID_LINK": "無効なリンクです",
-			"MOBILE_GUIDE_MODAL_DESC_1": "初期ログインの時に",
-			"MOBILE_GUIDE_MODAL_DESC_2": "PC環境でまずパスワードを再設定した後、",
-			"MOBILE_GUIDE_MODAL_DESC_3": "ご利用いただけます。",
-			"MOBILE_GUIDE_MODAL_TITLE": "モバイル環境では利用不可",
-			"RESET_PASSWORD": "パスワード再設定"
 		}
 	},
 	"AUTOMATION": {
@@ -940,6 +934,7 @@
 		"PROFILE": {
 			"ALT_E_UPDATE": "個人情報編集に失敗",
 			"ALT_S_UPDATE": "個人情報編集完了",
+			"CURRENT_PASSWORD": "",
 			"EMAIL": "メールアドレス",
 			"EMAIL_REQUIRED": "メールアドレスを入力してください",
 			"GROUP": "グループ/会社",
@@ -950,6 +945,7 @@
 			"MOBILE": "電話番号",
 			"NAME": "お名前",
 			"NAME_REQUIRED": "お名前を入力してください。",
+			"NEW_PASSWORD": "",
 			"PASSWORD": "パスワード",
 			"PASSWORD_CHECK": "パスワードの確認",
 			"PASSWORD_CHECK_INVALID": "同じパスワードを入力してください",
@@ -1628,6 +1624,14 @@
 				"ACCOUNT_N_PROFILE": "アカウント＆情報",
 				"BASE_INFORMATION": "基本情報",
 				"CHANGE_PASSWORD": "パスワード変更",
+				"NOTIFICATION_EMAIL": {
+					"CHANGE": "",
+					"HELP_TEXT": "",
+					"NOT_VERIFIED": "",
+					"TITLE": "",
+					"VERIFIED": "",
+					"VERIFY": ""
+				},
 				"SAVE_CHANGES": "変更事項を保存"
 			},
 			"API_KEY": {
