+++ packages/language-pack/ko.json
@@ -36,6 +36,12 @@
 				"TITLE": "",
 				"USER_ID": ""
 			},
+			"MOBILE": {
+				"MOBILE_GUIDE_MODAL_DESC_1": "",
+				"MOBILE_GUIDE_MODAL_DESC_2": "",
+				"MOBILE_GUIDE_MODAL_DESC_3": "",
+				"MOBILE_GUIDE_MODAL_TITLE": ""
+			},
 			"RESET": {
 				"CONFIRM_PASSWORD": "",
 				"DESC": "",
@@ -71,18 +77,6 @@
 				"RESET_PASSWORD": "비밀번호 재설정",
 				"TITLE": ""
 			}
-		},
-		"RESET_PASSWORD_PAGE": {
-			"ALREADY_RESET_TEXT": "패스워드가 이미 변경되었습니다. 변경된 패스워드로 로그인하세요.",
-			"GO_TO_LOGIN_PAGE": "로그인 페이지로 이동",
-			"GO_TO_LOGIN_WITHOUT_CHANGE": "변경하지 않고 로그인으로 돌아가기",
-			"HELP_TEXT": "다음 계정의 새로운 비밀번호 입력 필요: ",
-			"INVALID_LINK": "잘못된 링크입니다.",
-			"MOBILE_GUIDE_MODAL_DESC_1": "초기 진입 시, ",
-			"MOBILE_GUIDE_MODAL_DESC_2": "PC환경에서 먼저 비밀번호 재설정 후",
-			"MOBILE_GUIDE_MODAL_DESC_3": "이용하실 수 있습니다.",
-			"MOBILE_GUIDE_MODAL_TITLE": "모바일 환경 이용 불가",
-			"RESET_PASSWORD": "비밀번호 재설정"
 		}
 	},
 	"AUTOMATION": {
@@ -940,6 +934,7 @@
 		"PROFILE": {
 			"ALT_E_UPDATE": "개인 정보 수정 실패",
 			"ALT_S_UPDATE": "개인 정보 수정 완료",
+			"CURRENT_PASSWORD": "",
 			"EMAIL": "이메일",
 			"EMAIL_REQUIRED": "이메일을 입력해주세요.",
 			"GROUP": "소속",
@@ -950,6 +945,7 @@
 			"MOBILE": "전화번호",
 			"NAME": "이름 ",
 			"NAME_REQUIRED": "이름을 입력해 주세요.",
+			"NEW_PASSWORD": "",
 			"PASSWORD": "비밀번호",
 			"PASSWORD_CHECK": "비밀번호 확인",
 			"PASSWORD_CHECK_INVALID": "동일한 비밀번호를 입력해 주세요.",
@@ -1628,6 +1624,14 @@
 				"ACCOUNT_N_PROFILE": "계정 & 프로필",
 				"BASE_INFORMATION": "기본 정보",
 				"CHANGE_PASSWORD": "비밀번호 변경 ",
+				"NOTIFICATION_EMAIL": {
+					"CHANGE": "",
+					"HELP_TEXT": "",
+					"NOT_VERIFIED": "",
+					"TITLE": "",
+					"VERIFIED": "",
+					"VERIFY": ""
+				},
 				"SAVE_CHANGES": "변경사항 저장 "
 			},
 			"API_KEY": {
